registry/access_control/pyproject.toml
[project]
name = "feathr_rbac"
version = "0.0.1"
description = "Python Package with Feathr RBAC routines server"

authors = [
    #    { "name" = "Full Name" },
]
maintainers = [
    #    { "name" = "Full Name" },
]
classifiers = [
    "Operating System :: Unix",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10"
]
requires-python = ">=3.9"
dependencies = [
    "pymssql==2.2.7",
    "fastapi==0.88.0",
    "uvicorn==0.20.0",
    "pyjwt==2.8.0",
    "requests==2.31.0",
    "rsa==4.9",
    "sqlalchemy==1.4.46",
    "loguru==0.7.0",
    "cryptography==41.0.3"
]

[project.optional-dependencies]
test = [
    #    "pytest>=x.x",
    #    "pytest-cov>=x.x"
]


[build-system]
requires = [
    "setuptools",
    "wheel"
]
